+++ client/src/components/nav-components/MyAccountMenu.js
@@ -1,0 +1,20 @@
+import React from 'react';
+// import { Link } from 'react-router-dom';
+// import { useDispatch, useSelector } from 'react-redux';
+// import IosMenu from 'react-ionicons/lib/IosMenu';
+// import useServerAPI from '../../hooks/useServerAPI.js';
+
+const MyAccountMenu = () => {
+
+
+    return (
+        <div className='my-account-menu' onClick={() => setShowMenu(true)}>
+            <button className='button'>testing</button>
+        </div>
+
+    )
+}
+
+
+
+export { MyAccountMenu as default }